- Remove the search wrapper. - Make the discrete legend search the results by default (based on the given value).
